moved history functions
