+++ src/app/features/dashboard/utils/graph.utils.spec.ts
@@ -1,11 +1,9 @@
-import { addDays, eachDay, subDays } from 'date-fns';
+import { addDays, eachDay } from 'date-fns';
 
 import { HistoryGrouping } from '../../../shared/models';
 import { GraphDataItem } from '../models';
 
-import {
-    DEFAULT_KEY, filterGroupingsByDateRange, mapToGraphData, padDateGraphData
-} from './graph.utils';
+import { DEFAULT_KEY, mapToGraphData, padDateGraphData } from './graph.utils';
 
 describe('Graph Utils', () => {
   describe('mapToGraphData', () => {
@@ -66,162 +64,6 @@ describe('Graph Utils', () => {
 
         expect(result[0].value).toBe(totalTime);
       });
-    });
-  });
-
-  describe('filterGroupingsByDateRange', () => {
-    const start = new Date(2018, 3, 1);
-    const end = addDays(start, 6);
-    const inRange = addDays(start, 3);
-    const outOfRangeAhead = addDays(start, 15);
-    const outOfRangeBehind = subDays(start, 3);
-    const range = eachDay(start, end);
-    const game = 'some game';
-
-    it('Should filter history items outside of date range', () => {
-      const grouping = getHistoryGrouping(game, 0);
-      grouping.historyItems = [{
-        id: 'some id 4',
-        game,
-        platform: 'some platform',
-        startTime: outOfRangeAhead.getTime(),
-        endTime: outOfRangeAhead.getTime(),
-        dateRange: [
-          outOfRangeAhead,
-          outOfRangeAhead
-        ]
-      }, {
-        id: 'some id 3',
-        game,
-        platform: 'some platform',
-        startTime: inRange.getTime(),
-        endTime: inRange.getTime(),
-        dateRange: [
-          inRange,
-          inRange
-        ]
-      }, {
-        id: 'some id 2',
-        game,
-        platform: 'some platform',
-        startTime: start.getTime(),
-        endTime: start.getTime(),
-        dateRange: [
-          start,
-          start
-        ]
-      }, {
-        id: 'some id',
-        game,
-        platform: 'some other platform',
-        startTime: outOfRangeBehind.getTime(),
-        endTime: outOfRangeBehind.getTime(),
-        dateRange: [
-          outOfRangeBehind,
-          outOfRangeBehind
-        ]
-      }];
-
-      const result = filterGroupingsByDateRange([grouping], range);
-
-      expect(result).toEqual([{
-        key: game,
-        totalTime: 0,
-        historyItems: [{
-          id: 'some id 3',
-          game,
-          platform: 'some platform',
-          startTime: inRange.getTime(),
-          endTime: inRange.getTime(),
-          dateRange: [
-            inRange,
-            inRange
-          ]
-        }, {
-          id: 'some id 2',
-          game,
-          platform: 'some platform',
-          startTime: start.getTime(),
-          endTime: start.getTime(),
-          dateRange: [
-            start,
-            start
-          ]
-        }]
-      }]);
-    });
-
-    it('Should update total time', () => {
-      const grouping = getHistoryGrouping(game, 7000);
-      grouping.historyItems = [{
-        id: 'some id 4',
-        game,
-        platform: 'some platform',
-        startTime: outOfRangeAhead.getTime(),
-        endTime: outOfRangeAhead.getTime() + 2000,
-        dateRange: [
-          outOfRangeAhead,
-          new Date(outOfRangeAhead.getTime() + 2000)
-        ]
-      }, {
-        id: 'some id 3',
-        game,
-        platform: 'some platform',
-        startTime: inRange.getTime(),
-        endTime: inRange.getTime() + 1000,
-        dateRange: [
-          inRange,
-          new Date(inRange.getTime() + 1000)
-        ]
-      }, {
-        id: 'some id 2',
-        game,
-        platform: 'some platform',
-        startTime: start.getTime(),
-        endTime: start.getTime() + 1000,
-        dateRange: [
-          start,
-          new Date(start.getTime() + 1000)
-        ]
-      }, {
-        id: 'some id',
-        game,
-        platform: 'some other platform',
-        startTime: outOfRangeBehind.getTime(),
-        endTime: outOfRangeBehind.getTime() + 3000,
-        dateRange: [
-          outOfRangeBehind,
-          new Date(outOfRangeBehind.getTime() + 3000)
-        ]
-      }];
-
-      const result = filterGroupingsByDateRange([grouping], range);
-
-      expect(result).toEqual([{
-        key: game,
-        totalTime: 2, // time on history items is in milliseconds; total time is in seconds
-        historyItems: [{
-          id: 'some id 3',
-          game,
-          platform: 'some platform',
-          startTime: inRange.getTime(),
-          endTime: inRange.getTime() + 1000,
-          dateRange: [
-            inRange,
-            new Date(inRange.getTime() + 1000)
-          ]
-        }, {
-          id: 'some id 2',
-          game,
-          platform: 'some platform',
-          startTime: start.getTime(),
-          endTime: start.getTime() + 1000,
-          dateRange: [
-            start,
-            new Date(start.getTime() + 1000)
-          ]
-        }]
-      }]);
     });
   });
