+++ src/app/features/dashboard/utils/graph.utils.ts
@@ -3,9 +3,7 @@ import { isSameDay } from 'date-fns';
 import { HistoryGrouping } from '../../../shared/models';
 import { GraphDataItem } from '../models';
 
-import {
-    formatDate, getElapsedTimeInSeconds, isInDateRange
-} from '../../../shared/utils/date.utils';
+import { formatDate } from '../../../shared/utils/date.utils';
 
 export const DEFAULT_KEY = 'Uncategorized';
 
@@ -14,28 +12,6 @@ export const mapToGraphData = (groupings: HistoryGrouping[]): GraphDataItem[] =>
     name: grouping.key === '' ? DEFAULT_KEY : grouping.key,
     value: grouping.totalTime
   });
-};
-
-export const filterGroupingsByDateRange = (groupings: HistoryGrouping[], dateRange: Date[]): HistoryGrouping[] => {
-  const groupingsToGraph: HistoryGrouping[] = [];
-  groupings.forEach(grouping => {
-    const groupingHasHistoryItemInRange = grouping.historyItems.some(item => isInDateRange(item.dateRange[0], dateRange));
-    if (groupingHasHistoryItemInRange) {
-      const filtedGrouping = getFilteredGrouping(grouping, dateRange);
-      groupingsToGraph.push(filtedGrouping);
-    }
-  });
-  return groupingsToGraph;
-};
-
-const getFilteredGrouping = (grouping: HistoryGrouping, dateRange: Date[]): HistoryGrouping => {
-  const historyItems = grouping.historyItems.filter(item => isInDateRange(item.dateRange[0], dateRange));
-  const totalTime = historyItems.map(item => getElapsedTimeInSeconds(item.startTime, item.endTime)).reduce((a, b) => a + b, 0);
-  return {
-    key: grouping.key,
-    totalTime,
-    historyItems
-  };
 };
 
 export const padDateGraphData = (graphItems: GraphDataItem[], dateRange: Date[]): GraphDataItem[] => {
